Refactor structure of library methods, Smooth scrolling in ControlBar component
