+++ src/components/social/Social.tsx
@@ -1,4 +1,4 @@
-import { Social } from '@/interfaces';
+import { Social } from '@/lib/interfaces';
 import Image from 'next/image';
 import Link from 'next/link';
